requirements.txt
Django==4.2.16
djangorestframework==3.15.2
openai==1.60.0
python-decouple==3.8
psycopg2-binary==2.9.9
tiktoken==0.8.0
